requirements.txt
# Core Framework Dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# AI/ML Dependencies
tensorflow==2.15.0
transformers==4.35.2
torch==2.1.1
torchvision==0.16.1
pillow==10.1.0
numpy==1.24.3
scikit-learn==1.3.2

# Image Processing
opencv-python==4.8.1.78
matplotlib==3.8.2

# API and Web Dependencies
pydantic==2.5.0
requests==2.31.0
httpx==0.25.2

# Utility Dependencies
python-dotenv==1.0.0
pyyaml==6.0.1
tqdm==4.66.1

# Development Dependencies
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
flake8==6.1.0

# Optional: For enhanced performance
accelerate==0.25.0
optimum==1.15.0